refactor: updated login loading logic
