+++ src/app/Header/Header.js
@@ -1,8 +1,8 @@
-import Link from "next/link";
-import "./Header.css";
-import HeaderLoggedIn from "./HeaderLoggedIn";
+import Link from 'next/link';
+import './Header.css';
+import HeaderLoggedIn from './HeaderLoggedIn';
 
-export default function Header() {
+export default function Header({ userData }) {
   return (
     <header>
       <ul>
@@ -12,37 +12,37 @@ export default function Header() {
           </Link>
         </li>
         <li>
-          {" "}
+          {' '}
           <Link className="nav-item" href="/About">
             About Us
           </Link>
         </li>
         <li>
-          {" "}
+          {' '}
           <Link className="nav-item" href="/Contact">
             Contact
           </Link>
         </li>
         <li>
-          {" "}
+          {' '}
           <Link className="nav-item" href="/Products">
             Products
           </Link>
         </li>
         <li>
-          {" "}
+          {' '}
           <Link className="nav-item" href="/Blogs">
             Blogs
           </Link>
         </li>
         <li>
-          {" "}
+          {' '}
           <Link className="nav-item" href="/Profile">
             My Profile
           </Link>
         </li>
       </ul>
-      <HeaderLoggedIn />
+      <HeaderLoggedIn userData={userData} />
     </header>
   );
 }
