+++ src/app/Header/HeaderLoggedIn.js
@@ -1,35 +1,25 @@
-"use client";
+'use client';
 
-import { useEffect, useState } from "react";
-import { getCurrentAuthUser } from "../_Services/authService";
-
-export default function HeaderLoggedIn() {
-  const [userData, setUserData] = useState(null);
-
-  useEffect(() => {
-    const fetchUserData = async () => {
-      const data = await getCurrentAuthUser();
-      setUserData(data);
-    };
-    fetchUserData();
-  }, []);
-
+export default function HeaderLoggedIn({ userData }) {
+  // Function to handle user logout
   function handleLogout() {
-    document.cookie = "token=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/;";
-    setUserData(null);
-    window.location.href = "/Login";
+    document.cookie = 'token=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/;';
+    window.location.href = '/Login'; // Redirect to login page
   }
 
-  if (userData) {
-    return (
-      <div className="header-loggedIn">
-        <span className="header-hello">{`Hello ${userData.firstName}!`} </span>
-        <span className="header-logout">
-          <button className="btn-logout" onClick={handleLogout}>
-            Log Out
-          </button>
-        </span>
-      </div>
-    );
+  // Ensure the header only renders if userData exists
+  if (!userData) {
+    return null;
   }
+
+  return (
+    <div className="header-loggedIn">
+      <span className="header-hello">{`Hello ${userData.firstName}!`} </span>
+      <span className="header-logout">
+        <button className="btn-logout" onClick={handleLogout}>
+          Log Out
+        </button>
+      </span>
+    </div>
+  );
 }
